fix layout, Variant.model

+++ ege/app/src/main/java/com/deadely/ege/model/Model.kt
@@ -16,7 +16,7 @@ data class University(
 data class Speciality(
     val _id: String,
     val name: String,
-    val point: Int,
+    val point: Double,
 ) : Parcelable
 
 //////////////////////////////
